Multi-part commit. #more
The home page now has some logic to redirect the user based on their
user-role. This could be made a lot more elegant than it is now, but
that can come later.

The admin UI groundwork is complete. The web side of file uploading is
complete... it just needs the stub to be filled out to process the
uploaded CSV file.